% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_column_names.R
\name{fix_column_names}
\alias{fix_column_names}
\title{fix_column_names}
\usage{
fix_column_names(x, char = "_")
}
\arguments{
\item{x}{a character vector}

\item{char}{replace special characters with this symbol}
}
\description{
removes special chars from names
}
