% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_column_names.R
\name{fix_names}
\alias{fix_names}
\title{fix_names}
\usage{
fix_names(x, char = "_")
}
\arguments{
\item{x}{a vector of column names}

\item{char}{substitute special char with this.}
}
\description{
Replace special characters in column names of a data.frame
}
\seealso{
make.names
}
