% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conf.R
\name{load_conf}
\alias{load_conf}
\title{Loading configuration files.}
\usage{
load_conf(x, check = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{x}{path to a configuration file}

\item{check}{in case of a configuration file, whether to check if files defined in parameters exists..}

\item{verbose}{be chatty ?}

\item{...}{Not used}
}
\description{
Read a tab delimted file and load them as options using \link{set_opts}
}
\details{
params ending in: "path$|dir$|exe$" are checked in case of check=TRUE.
}
\seealso{
\link{get_opts} \link{set_opts} \link{print.opts} \link{read_sheet}
}

