% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/opts.R
\name{get_opts}
\alias{get_opts}
\alias{print.opts}
\alias{set_opts}
\title{Setting/loading and extracting various options into the environment}
\usage{
get_opts(x, envir = opts)

set_opts(x, envir = opts)

\method{print}{opts}(x, ...)
}
\arguments{
\item{x}{get_opts(): a character vector of names of options to extract.
set_opts(): a named list with all the options to be set.}

\item{envir}{environ used to store objects}

\item{...}{not used}
}
\description{
load_conf:
This function extracts the options set by functions like set_opts/load_conf

print pkg options as a pretty table
}
\details{
In case of print.opts(), param ... is passed onto print.
}
\examples{
get_opts()
get_opts("flow_run_path")
## set _opts
set_opts(list(flow_run_path = "~/mypath"))
}
\seealso{
\link{load_conf}
}

