% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_opts.R
\name{parse_opts}
\alias{parse_opts}
\title{Parse options to expand \code{{{variable}}} into their respective values}
\usage{
parse_opts(lst, envir)
}
\arguments{
\item{lst}{a list of configuration options to parse}

\item{envir}{environ used to store objects. Default is a environ object called opts [params:::opts]}
}
\description{
This function is internally called by \link{set_opts} and \link{load_opts}
}
